version 0.2

+++ app/src/main/java/rdp/fit/bstu/application/MainActivityThird.java
@@ -19,6 +19,11 @@ public class MainActivityThird extends AppCompatActivity {
     private String result = null;
     private TextView textView = null;
 
+    // Метод проверки, существует ли файл и не является ли он каталогом
+    public static boolean isFileExists(File file) {
+        return file.exists() && !file.isDirectory();
+    }
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         try {
@@ -40,17 +45,21 @@ public class MainActivityThird extends AppCompatActivity {
             textView.setText(result);
             FileWriter fWriter;
             File sdCardFile = new File(Environment.getExternalStorageDirectory() + "\\filename.txt");
-            JSONObject object = new JSONObject();
-            object.put("NAME", intent.getStringExtra(MainActivity.EXTRA_NAME));
-            object.put("SURNAME", intent.getStringExtra(MainActivity.EXTRA_SURNAME));
-            object.put("MARRIED", intent.getStringExtra(MainActivity.EXTRA_MARRIED) == "true" ? "Married" : "Not Married");
-            object.put("DATE", intent.getStringExtra(MainActivitySecond.EXTRA_DATE));
-            object.put("PASSPORTNUMBER", intent.getStringExtra(MainActivitySecond.EXTRA_PASSPORTNUMBER));
-            object.put("SEX", intent.getStringExtra(MainActivitySecond.EXTRA_SEX));
-            fWriter = new FileWriter(sdCardFile, true);
-            fWriter.write(object.toString());
-            fWriter.flush();
-            fWriter.close();
+            if(!isFileExists(sdCardFile)){
+                JSONObject object = new JSONObject();
+                object.put("NAME", intent.getStringExtra(MainActivity.EXTRA_NAME));
+                object.put("SURNAME", intent.getStringExtra(MainActivity.EXTRA_SURNAME));
+                object.put("MARRIED", intent.getStringExtra(MainActivity.EXTRA_MARRIED) == "true" ? "Married" : "Not Married");
+                object.put("DATE", intent.getStringExtra(MainActivitySecond.EXTRA_DATE));
+                object.put("PASSPORTNUMBER", intent.getStringExtra(MainActivitySecond.EXTRA_PASSPORTNUMBER));
+                object.put("SEX", intent.getStringExtra(MainActivitySecond.EXTRA_SEX));
+                fWriter = new FileWriter(sdCardFile, true);
+                fWriter.write(object.toString());
+                fWriter.flush();
+                fWriter.close();
+                return;
+            }
+
         }catch (Exception exception){
             System.out.println("MainActivityThird\t"+exception.getMessage());
             textView.setText(exception.getMessage());
